Remove Make A Language series
This lives in a separate repo now (arzg/lang). Note that all URLs stay
the same, so this change doesn’t break any links. I did this because I
wanted to play around with the design of my site and that of the
language tutorials, but felt that e.g. a style that would fit a
personal blog might not necessarily fit a long (59,072 words) series
of technical blog posts. Regardless of if I ever do mess around with
the styling, I think it’s nice conceptually to have them a little more
separate.